requirements.txt
django==5.2.4
django-allauth==65.9.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1 
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0

